All -> Delete, Move and rename some components
