--- src/components/Building.tsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const Building = () => {
-    return (
-        <h1 className='mt-80 text-[200px] leading-none font-extrabold text-text uppercase'>En construction</h1>
-    )
-}
-
-export default Building
