+++ src/components/Carousel/Dot.tsx
